New changes request added
Thank for you for making the changes.

Could we add two more things under the satellite dropdown? Add a box for
“Window Time.” While I know you already have a TIME location, that works
well for live feeds with Hot Channel and TVU grid.

Also next to the addition under satellite can we add a box for
Troubleshoot #? We can ask folks to include contact in case there is a
problem with a feed.

The change you made from ‘add’ to ‘submit’ was a good idea. Tks

Lastly, under format can we add a ‘Live’ distinction?